update reorganize code structure, add utils py script
